Grouping training and training-tutorial material
The tutorial, chapters and images are now under the same repo as all the
activities and presentation-relates files. And it will be arranged to be
rendered in a different place then before. This files were originally in
the repo https://github.com/bpiereck/training-material_VIBTechTraining/tree/master/topics/git-introduction/tutorials