Course information update
The front page file summarizing the information about this course, has now the iformation that
highligts thecontributrs, the program and links to the 'Get ready' session so trainers and trainees
can prepare their computers and have the pre-requisites(shell cmd) fufilled. It is still missing the
about sessoin as much as the learning outcomes and licesce for material re-use.